backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django-cors-headers==3.13.0
django==4.0.5
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
pyjwt==2.4.0; python_version >= '3.6'
pytz==2022.1
sqlparse==0.4.2; python_version >= '3.5'
